requirements.txt
celery==5.2.7
channels==3.0.4
django==4.1.5
django-celery-beat==2.4.0
django-elasticsearch-dsl==7.3
django-neomodel==0.0.7
django-social-share==2.3.0
django-taggit==3.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
django-redis==5.4.0